moving eventlib patching to Click
