+++ celery/__main__.py
@@ -2,15 +2,12 @@
 
 import sys
 
-# from . import maybe_patch_concurrency
 
 __all__ = ('main',)
 
 
 def main():
     """Entrypoint to the ``celery`` umbrella command."""
-    # if 'multi' not in sys.argv:
-    #     maybe_patch_concurrency()
     from celery.bin.celery import main as _main
     sys.exit(_main())
